% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_parser}
\alias{find_parser}
\title{Attempt to find the correct date parser for a vector of character dates}
\usage{
find_parser(x)
}
\arguments{
\item{x}{A list or character vector of dates to analyse}
}
\value{
The function to use for parsing
}
\description{
Check 5% of all rows in a date column so as to find the correct
  order for the parser (ymd, dmy, mdy, etc.). Then we can just use that instead
  of a more permissive but potentially slower parser.
}
