% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_df_coltypes}
\alias{get_df_coltypes}
\title{Get \code{tibble}-guessed data types for each dataframe column}
\usage{
get_df_coltypes(df)
}
\arguments{
\item{df}{Dataframe created from the loaded CSV file.}
}
\value{
A list where each name is the \code{inputId} of an input in the
  UI, and each value is a vector of column names with the correct
  characteristics.
  @import magrittr
  @import dplyr
}
\description{
Get all of \code{df} columns' data types and put them in
  a list usable to update the UI
}
