% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_mappings_inputs}
\alias{update_mappings_inputs}
\title{Update variables drop-downs in the charts UI}
\usage{
update_mappings_inputs(df, session)
}
\arguments{
\item{df}{Dataframe created from the loaded CSV file.}

\item{session}{The session object passed to function given to shinyServer.}
}
\value{
Nothing, we only want a side effect here
}
\description{
When a dataframe has been type-checked from the loaded CSV
  file, update the select list aesthetics mappings controls on the charts UI tab.
}
