% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_type_inputs}
\alias{update_type_inputs}
\title{Update type controls in the UI}
\usage{
update_type_inputs(df, session)
}
\arguments{
\item{df}{Dataframe created from the loaded CSV file.}

\item{session}{The session object passed to function given to shinyServer.}
}
\value{
Nothing, we only want a side effect here
}
\description{
When a new CSV file is loaded, update the select list input controls
  on the starting UI tab
}
